--- client/src/component/Home/Home.module.css
@@ -1,7 +1,0 @@
-.page{
-    background-color: rgba(43,38,38);
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-around;
-}
-
