+++ client/src/component/ProductDetail/CardDetailProduct.jsx
@@ -1,12 +1,12 @@
 import React, { useState, useEffect } from 'react'
-import Style from './CardProduct.module.css'
+import Style from './CardDetailProduct.module.css'
 import { AiFillStar } from 'react-icons/ai'
 import { IconContext } from "react-icons";
-import productJson from '../Home/Products.json'
+import productJson from '../Products/Products.json'
 import { useParams } from 'react-router-dom';
 
 
-const CardProduct = () => {
+const CardDetailProduct = () => {
 
   const paramsId = useParams()
   let productFiltered;
@@ -18,20 +18,6 @@ const CardProduct = () => {
   })
 
   console.log(productFiltered.name)
-
-
-  // for (let i = 0; i < productJson.length; i++) {
-  //     if(productJson[i].id === paramsId.id){
-  //     productsId = productJson[i]
-  //     return productsId;
-  //   }
-  //   }
-
-
-
-
-
-
 
   const [counter, setCounter] = useState(0);
 
@@ -86,4 +72,4 @@ const CardProduct = () => {
   )
 }
 
-export default CardProduct
+export default CardDetailProduct
