+++ client/src/component/Products/Products.jsx
@@ -1,22 +1,33 @@
 import React from 'react'
-import style from './Recipes.module.css'
+import style from "./Products.module.css"
+import Card from "./CardProducts/Card"
+import Product from "./Products.json"
+import {Link} from 'react-router-dom'
 
 
-export default function Products(props){
-    return (
-        <div className={style.card}>
-       
-        <img src={props.img} alt={props.name}  />
-           
-        <h1>{props.name}</h1>
+const Products = () => {
+
+  return (
+    <div>
+        <div className={style.page}> 
         
-            <div className="flex-auto flex space-x-4">
-            <button className={style.button}>Ver Más</button>
-           
-           
-            </div>
-        
+        {
+        Product.map((e)=>{
+            return(
+              <Link to={`/productDetail/${e.id}`}>
+                <Card 
+                  key={e.name}
+                  name= {e.name}
+                  img= {e?.image}
+                  precio= {e.price}/>
+              </Link> 
+            )
+            })
+        }
         </div>
-
-    )
+      </div>
+  )
 }
+
+export default Products
+
